Fix test errors
Moved tests from transport to ref.dart since decode has been moved out of transport.
