+++ packages/firebase_data_connect/firebase_data_connect/test/src/network/rest_transport_test.dart
@@ -311,203 +311,5 @@ void main() {
         ),
       ).called(1);
     });
-    test('invokeOperation should throw an error if the server throws one',
-        () async {
-      final mockResponse = http.Response(
-        '''
-{
-    "data": {},
-    "errors": [
-        {
-            "message": "SQL query error: pq: duplicate key value violates unique constraint movie_pkey",
-            "locations": [],
-            "path": [
-                "the_matrix"
-            ],
-            "extensions": null
-        }
-    ]
-}''',
-        200,
-      );
-      when(
-        mockHttpClient.post(
-          any,
-          headers: anyNamed('headers'),
-          body: anyNamed('body'),
-        ),
-      ).thenAnswer((_) async => mockResponse);
-
-      final deserializer = (String data) => 'Deserialized Data';
-
-      expect(
-        () => transport.invokeOperation(
-          'testQuery',
-          'executeQuery',
-          deserializer,
-          null,
-          null,
-          null,
-        ),
-        throwsA(isA<DataConnectError>()),
-      );
-    });
-    test('invokeOperation should decode a partial error if available',
-        () async {
-      final mockResponse = http.Response(
-        '''
-        {
-            "data": {"abc": "def"},
-            "errors": [
-                {
-                    "message": "SQL query error: pq: duplicate key value violates unique constraint movie_pkey",
-                    "locations": [],
-                    "path": [
-                        "the_matrix"
-                    ],
-                    "extensions": null
-                }
-            ]
-        }''',
-        200,
-      );
-      when(
-        mockHttpClient.post(
-          any,
-          headers: anyNamed('headers'),
-          body: anyNamed('body'),
-        ),
-      ).thenAnswer((_) async => mockResponse);
-
-      final deserializer = (String data) {
-        Map<String, dynamic> decoded = jsonDecode(data) as Map<String, dynamic>;
-        return AbcHolder(decoded['abc']!);
-      };
-
-      expect(
-        () => transport.invokeOperation(
-          'testQuery',
-          'executeQuery',
-          deserializer,
-          null,
-          null,
-          null,
-        ),
-        throwsA(predicate((e) =>
-            e is DataConnectOperationError &&
-            e.response.rawData!['abc'] == 'def' &&
-            e.response.errors.first.message ==
-                'SQL query error: pq: duplicate key value violates unique constraint movie_pkey' &&
-            (e.response.errors.first.path[0] as DataConnectFieldPathSegment)
-                    .field ==
-                'the_matrix' &&
-            e.response.data is AbcHolder &&
-            (e.response.data as AbcHolder).abc == 'def')),
-      );
-    });
-    test(
-        'invokeOperation should decode a partial error if no path is specified',
-        () async {
-      final mockResponse = http.Response(
-        '''
-        {
-            "data": {"abc": "def"},
-            "errors": [
-                {
-                    "message": "invalid pkey",
-                    "locations": [],
-                    "path": null,
-                    "extensions": null
-                }
-            ]
-        }''',
-        200,
-      );
-      when(
-        mockHttpClient.post(
-          any,
-          headers: anyNamed('headers'),
-          body: anyNamed('body'),
-        ),
-      ).thenAnswer((_) async => mockResponse);
-
-      final deserializer = (String data) {
-        Map<String, dynamic> decoded = jsonDecode(data) as Map<String, dynamic>;
-        return AbcHolder(decoded['abc']!);
-      };
-
-      expect(
-        () => transport.invokeOperation(
-          'testQuery',
-          'executeQuery',
-          deserializer,
-          null,
-          null,
-          null,
-        ),
-        throwsA(predicate((e) =>
-            e is DataConnectOperationError &&
-            e.response.rawData!['abc'] == 'def' &&
-            e.response.errors.first.message == 'invalid pkey' &&
-            e.response.errors.first.path.isEmpty &&
-            e.response.data is AbcHolder &&
-            (e.response.data as AbcHolder).abc == 'def')),
-      );
-    });
-    test(
-        'invokeOperation should decode a partial error if list path is specified',
-        () async {
-      final mockResponse = http.Response(
-        '''
-        {
-            "data": {"abc": "def"},
-            "errors": [
-                {
-                    "message": "invalid pkey",
-                    "locations": [],
-                    "path": [1,2,3],
-                    "extensions": null
-                }
-            ]
-        }''',
-        200,
-      );
-      when(
-        mockHttpClient.post(
-          any,
-          headers: anyNamed('headers'),
-          body: anyNamed('body'),
-        ),
-      ).thenAnswer((_) async => mockResponse);
-
-      final deserializer = (String data) {
-        Map<String, dynamic> decoded = jsonDecode(data) as Map<String, dynamic>;
-        return AbcHolder(decoded['abc']!);
-      };
-
-      expect(
-          () => transport.invokeOperation(
-                'testQuery',
-                'executeQuery',
-                deserializer,
-                null,
-                null,
-                null,
-              ), throwsA(predicate((e) {
-        return e is DataConnectOperationError &&
-            e.response.rawData!['abc'] == 'def' &&
-            e.response.errors.first.message == 'invalid pkey' &&
-            e.response.errors.first.path.length == 3 &&
-            e.response.errors.first.path.first
-                is DataConnectListIndexPathSegment &&
-            e.response.data is AbcHolder &&
-            (e.response.data as AbcHolder).abc == 'def';
-      })));
-    });
   });
-}
-
-class AbcHolder {
-  String abc;
-  AbcHolder(this.abc);
 }
